Use app context provider

+++ src/components/ChabloomTable/Heading.tsx
@@ -1,21 +1,17 @@
 import * as React from "react";
 import { NavLink } from "react-router-dom";
 
-import { User } from "oidc-client";
-
 import { ButtonGroup, IconButton, lighten, LinearProgress, Toolbar, Tooltip, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import { Cancel, Delete, Edit, FilterList, Receipt, Save, Schedule } from "@material-ui/icons";
 import { Alert, AlertTitle } from "@material-ui/lab";
 
-import { BaseApiType, BaseViewModel, TenantViewModel } from "../../types";
+import { BaseApiType, BaseViewModel } from "../../types";
 
 import { ChabloomTableColumn } from "./Column";
 
 interface Props {
-  user: User | undefined;
-  tenant: TenantViewModel | undefined;
-  api: BaseApiType<BaseViewModel>;
+  api: BaseApiType<BaseViewModel> | undefined;
   title: string;
   columns: Array<ChabloomTableColumn>;
   methods: Array<"add" | "edit" | "delete" | "payment" | "paymentSchedule">;
@@ -60,50 +56,56 @@ const useStyles = makeStyles((theme) => ({
 const ChabloomTableActionButtons: React.FC<Props> = (props) => {
   const addItem = async () => {
     props.setProcessing(true);
-    const [newItem, err] = await props.api.addItem(props.editItem);
-    if (!err) {
-      props.setData([
-        ...props.data.slice(0, props.selectedIndex),
-        { ...newItem },
-        ...props.data.slice(props.selectedIndex + 1),
-      ]);
-      props.setSelectedIndex(-1);
-      props.setEditIndex(-1);
-      props.setEditItem({} as BaseViewModel);
-      props.setAdding(false);
-      props.setError("");
-    } else {
-      props.setError(err);
+    if (props.api) {
+      const [newItem, err] = await props.api.addItem(props.editItem);
+      if (!err) {
+        props.setData([
+          ...props.data.slice(0, props.selectedIndex),
+          { ...newItem },
+          ...props.data.slice(props.selectedIndex + 1),
+        ]);
+        props.setSelectedIndex(-1);
+        props.setEditIndex(-1);
+        props.setEditItem({} as BaseViewModel);
+        props.setAdding(false);
+        props.setError("");
+      } else {
+        props.setError(err);
+      }
     }
     props.setProcessing(false);
   };
   const editItem = async () => {
     props.setProcessing(true);
-    const [newItem, err] = await props.api.editItem(props.editItem);
-    if (!err) {
-      props.setData([
-        ...props.data.slice(0, props.selectedIndex),
-        { ...newItem },
-        ...props.data.slice(props.selectedIndex + 1),
-      ]);
-      props.setSelectedIndex(-1);
-      props.setEditIndex(-1);
-      props.setEditItem({} as BaseViewModel);
-      props.setError("");
-    } else {
-      props.setError(err);
+    if (props.api) {
+      const [newItem, err] = await props.api.editItem(props.editItem);
+      if (!err) {
+        props.setData([
+          ...props.data.slice(0, props.selectedIndex),
+          { ...newItem },
+          ...props.data.slice(props.selectedIndex + 1),
+        ]);
+        props.setSelectedIndex(-1);
+        props.setEditIndex(-1);
+        props.setEditItem({} as BaseViewModel);
+        props.setError("");
+      } else {
+        props.setError(err);
+      }
     }
     props.setProcessing(false);
   };
   const deleteItem = async () => {
     props.setProcessing(true);
-    const err = await props.api.deleteItem(props.editItem);
-    if (!err) {
-      props.setData([...props.data.slice(0, props.selectedIndex), ...props.data.slice(props.selectedIndex + 1)]);
-      props.setSelectedIndex(-1);
-      props.setDeleteIndex(-1);
-    } else {
-      props.setError(err);
+    if (props.api) {
+      const err = await props.api.deleteItem(props.editItem);
+      if (!err) {
+        props.setData([...props.data.slice(0, props.selectedIndex), ...props.data.slice(props.selectedIndex + 1)]);
+        props.setSelectedIndex(-1);
+        props.setDeleteIndex(-1);
+      } else {
+        props.setError(err);
+      }
     }
     props.setProcessing(false);
   };
